requirements.txt
django
django-tastypie
psycopg2
lxml
defusedxml
